[1/2] SystemUI: Add time-context headers to the notification header
	[2/2] in packages/apps/Settings

This commit adds image headers to the notification panel header,
based on the date/time.

There is a default provider that provides sky pictures
as well as special events like Christmas and new
year's eve.

More will be coming, with eventually a way to use an external
(as in, Internet) provider for those pictures.

Signed-off-by: OwnDroid <askowndroid@gmail.com>